add website_iframe_host module
