+++ main/local_modules/frontend_base/controllers/base.py
@@ -57,10 +57,7 @@ def small_image_url(record, field):
                 sudo_record.write_date or sudo_record.create_date or ''
             ).hexdigest()[0:7]
         )
-        record.small_image_url = '/website/image/%s/%s/%s' % (
-        model, id_, field)
-        # else:
-        # _logger.debug('Great found small image url for %s!', record.id)
+        record.small_image_url = '/website/image/%s/%s/%s' % (model, id_, field)
 
     return record.small_image_url
