+++ main/local_modules/website_iframe_host/README.rst
@@ -1,0 +1,11 @@
+Website Iframe Host
+===================
+We were asked to embed a part of our website in an iframe. The iframe should be
+restricted to a given hostname and should have a given filter.
+In the request, it was to embed the map focused on France.
+
+The embed map should not have the header with the link to the other pages.
+
+Contributors
+------------
+* Jordi Riera <kender.jr@gmail.com>
