+++ main/local_modules/website_iframe_host/__openerp__.py
@@ -20,24 +20,19 @@
 ##############################################################################
 
 {
-    'name': 'Frontend Base',
+    'name': 'Website Iframe Host',
     'version': 'beta',
     'author': 'cgstudiomap',
     'maintainer': 'cgstudiomap',
     'license': 'AGPL-3',
     'category': 'Web',
-    'summary': 'Base for frontend pages.',
+    'summary': 'Allow another website to host an iframe from us.',
     'depends': [
         'website',
-        'website_iframe_host',
-        'res_group_archetype',
-        'res_partner_industry',
     ],
     'data': [
-        'templates/template_body.xml',
-        'templates/template_engine.xml',
-        'templates/template_head.xml',
-        'templates/template_html.xml',
+        'security/ir.model.access.csv',
+        'views/website_iframe_host.xml',
     ],
     'installable': True,
 }
