+++ main/local_modules/website_iframe_host/models/__init__.py
@@ -19,25 +19,4 @@
 #
 ##############################################################################
 
-{
-    'name': 'Frontend Base',
-    'version': 'beta',
-    'author': 'cgstudiomap',
-    'maintainer': 'cgstudiomap',
-    'license': 'AGPL-3',
-    'category': 'Web',
-    'summary': 'Base for frontend pages.',
-    'depends': [
-        'website',
-        'website_iframe_host',
-        'res_group_archetype',
-        'res_partner_industry',
-    ],
-    'data': [
-        'templates/template_body.xml',
-        'templates/template_engine.xml',
-        'templates/template_head.xml',
-        'templates/template_html.xml',
-    ],
-    'installable': True,
-}
+from . import website
