+++ main/local_modules/website_iframe_host/models/website.py
@@ -1,0 +1,13 @@
+from openerp import models, fields
+
+
+class WebsiteIframeHost(models.Model):
+    """Represent the name of a host that can embed our listing."""
+    _name = 'website.iframe.host'
+    _description = _name
+
+    host = fields.Char('Host', help='Name of the the host that will embed the listing.')
+    search_domain = fields.Char(
+        'Search Domain',
+        help='Domain that will be injected in searches for the given host.'
+    )
